vCalendar 1.0 TZ parsing: now searches in internal zone list
Before, a TZ without a subsequent DAYLIGHT in a vCalendar 1.0
was not attempted to convert into a known time zone from the
internal list, but was left a plain minute offset.
This fallback type of internal time zone representation is
not desirable, so we now try to find a matching internal
time zone even if no DAYLIGHT property is found.

For example, UTC is detected as such from a TZ:+00:00 now.