mime-dir: added "expandedlist" option for <property values="..">
The "values" attribute in <property> can now be set
to "expandedlist" (instead of just "list") for
properties that consist of an open list of values.

This new opetion causes such properties to be rendered
as a series of occurrences of the property representing
each a single value of the list, like:

  EXDATE:20090511T211133
  EXDATE:20090611T211133
  EXDATE:20090711T211133

instead of

  EXDATE:20090511T211133,20090611T211133,20090711T211133

It does not affect parsing, where both versions (even
when mixed in the same record) are allowed.

This option was primarily added for Scheduleworld
compatibiliy which seems to have problems with the
list form.