Scripting: Made SETLOG, SETDEBUGLOG and SETREADONLY generally accessible
up to now, these three functions were available in the
<loginXXXXscript> scripts only. Now they are generally
available in all scripts.

To allow switching debug logs dynamically in <sessioninitscript> from
a session variable defined from the app level, the session context in
non-initial calls to InternalResetSessionEx() is now retained.

Before, the session-level script context (fSessionScriptContextP) was
cleared and re-built each time InternalResetSessionEx() was called
so far.
Note that the <sessioninitscript> is called each time
InternalResetSessionEx() runs, which is once at the very beginning
of the session and in case of client sessions, once later at
SelectProfile().