client: added "smart" retry (mainly as ZYB workaround)
If the new <smartauthretry> client setting is on (=default),
the client will change the retry behaviour as defined with
newSessionForAuthRetry and noRespURIForAuthRetry to the
opposite after MAX_NORMAL_AUTH_RETRIES have failed.

So if the standard behaviour (as used in all Synthesis
production clients) of doing the retry in a completely
new session does not work, the client will now add
some retries with doing the retry in the same session.

This helps with servers like ZYB which leave the session
running and require the retry to address the already
started session (via RespURI) to correctly auth with
the nonce sent.

It also works the other way around - if standard auth
retry is set to retry in-session, smart retry will
do a fresh-session retry after MAX_NORMAL_AUTH_RETRIES
failed in-session retries.

So this "smart" auth should now make the client succeed
logging in with any server (those with the ZYB problem,
and those with the opposite problem, like the by now
probably irrelevant Magically)