engine 3.2.0.32: cleaned up binfile changelog update tracking
NOTE: UPDATING TO THIS VERSION WILL CAUSE A SLOW SYNC because
changelogs need to be reset.


Several changes and fixes to get two-level changelog handling in binfile right
when used with API DB (that is, with a DB which detects changes related to a
token passed in StartDataRead()).

The old implementation confused the meaning of the tokens
(and cmpRef dates) towards the DB API with the meaning of the
internal modCount anchors.

With BASED_ON_BINFILE_CLIENT, binfileXXX maintains the changelog
which provides the last-two-way-sync and last-suspend points in
time to select records from the database.
Towards the DB API, all binfileXXX needs to know are the
CHANGES SINCE LAST UPDATE OF THE CHANGELOG. So only ONE token
must be passed at StartDataRead, representing the time of last
changelog update.

In customimplXXX and its derivates, this means that fPreviousToRemoteSyncCmpRef
and fPreviousToRemoteSyncIdentifier get a new meaning and represent the
time of last changelog update. fPreviousSuspendCmpRef and fPreviousSuspendIdentifier
are not used (but set equal to fPreviousToRemoteSyncXXXX) when engine
is built as BASED_ON_BINFILE_CLIENT.

This means that database plugins that do NOT need to run with
non-BASED_ON_BINFILE_CLIENT engines, also don't need to implement
change detection relative to two tokens.
Just being able to list the changes since the time represented by ONE
token (first in StartDataRead) is sufficient.

Changes done:

- The comparison timestamp and identifier (token) are now saved
  in the changelog header itself, because there is ONE changelog
  across all profiles.

- lastChangeCheck and dummyIdentifier1 in the target are no longer used,
  but still updated for debug purposes.

- Added <cmptimestampatend> option to specify when the reference
  time for last changelog update should be sampled. For DBs that
  cannot put the sync-touched records to the start-of-sync
  timestamp, but can guarantee no DB changes occur DURING sync,
  this option can be set. It is similar to <synctimestampatend>
  (which does not exist for BASED_ON_BINFILE_CLIENT builds).
  <cmptimestampatend> defaults to defined(SYNCTIME_IS_ENDOFSESSION)
  to keep compatibility with older targets which define this.
  For SDK client engines and libsynthesis the default is FALSE.

- made sure failure to allocate changelog memory returns LOCERR_OUTOFMEM,
  not only LOCERR_UNDEFINED. A memory error allocating the changelog is
  something fatal, but it can happen under normal circumstances.
  LOCERR_UNDEFINED is for really dubious error conditions...