logging: use same TDebugLogger for all active engines and the session
The goal of this patch is to use just one logger instance for all
debug logging in SyncEvolution:
http://bugzilla.moblin.org/show_bug.cgi?id=3474

SyncEvolution chooses a unique log directory and never runs more than
one session inside it without clearing the directory first, therefore
having sysync_linux.html and a separate session .html with a varying
file name is overkill and a nuisance.

The approach taken in this patch is that normally, as before, a
TSyncAppBase creates and owns its own logger. If the GUI decides that
it wants a shared logger, it has to call shareLogger(). This
information is later picked up by further engine and session
instances.

The API to access shareLogger() and the whole approach are tentative,
so the minimum number of changes were made to let a client app reach
that call. In particular, TEngineModuleBridge doesn't support it yet.

SyncEvolution currently instantiates two engines. The
purpose of the first one is solely to provide a logger. No DB plugins
are configured. For this to work the check for configured datastores
had to be disabled. The second one is properly instantiated with
all all datastores.

The debug options of the second engine are ignore, because installing
them would reset the log file.