engine logging: new API with more information provided to the engine
The SySyncDebugPuts() call is tentative and not supported by the
TEngineBridge. It has to be called directly. Not all of the information
is currently used. To change that, TClientEngineInterface::debugPuts()
must pass them to an enhanced TDebugLogger implementation.

The main purpose right now is that SyncEvolution has different log
levels, which cannot be properly represented in the Synthesis log
without the possibility to specify flags.