autotools: add --enable-evolution-compatibility
The purpose of this switch is to build a libsynthesis.so which
does not depend on a specific release of libecal or libical. Such
a dependency is problematic because it limits the compatibility
of the library with different Linux distributions (libecal is
often different among those) and forces applications to use
the same libical library as the library.

The trick used by this patch is to look up the required functions
dynamically in the namespace of the application. As long as the
prototype of the four ical functions doesn't change (which is
unlikely) this works correctly. The code also takes into account
that in some Evolution releases the semantic of icalcomponent_as_ical_string()
was temporarily changed so that the caller has to free the result;
those libs export ical_memfixes.

For details, see http://bugzilla.moblin.org/show_bug.cgi?id=4441
"minimize external dependencies of non-Moblin binaries"