client + statistics workaround: treat status 200 as 201 for Replace cmd (which is really an Add)
synthesis always send replace cmd for a update or add; if this is a real add
command, servers (such as google, zyb) may return 200 (OK) instead of 201
(Added). This cause synthesis statics is wrong, count server add as update.
Workaround this by treating status 200 as status 201 for a Replace Cmd (which is
really a Add cmd).