autotools: added synthesis.pc file for pkg-config
This patch adds the normal synthesis.pc file, so that other programs
using pkg-config will automatically pick up the right paths and options.
The file assumes that users of libsynthesis will link against both
libsynthesissdk.a and libsynthesis.so, like SyncEvolution does. The
advantage is that the dependency is visible to packaging tools.