slow sync: avoid empty anchors, that confuses ScheduleWorld
This patch fixes http://bugzilla.moblin.org/show_bug.cgi?id=4703,
"ScheduleWorld: delete on server can not sync to client (after doing slow sync)".

Somehow the ScheduleWorld server has problems with <last></last>.
Saving the mapping on the server fails in that case. Mark called
an empty anchor string invalid.

I don't agree with that and think this issue should be fixed on
the server, but as it can be worked around easily (and perhaps
more quickly) on the client side, here's the patch for it.

Instead of "" it now uses "no local anchor". For symmetry reasons
(not required to solve this problem) the last remote anchor is
set to "no remote anchor".