mimedirprofile: autoenddate 23:59:59 detection bug fixed
to detect 23:59:59, the timestamp was converted to item time
zone (which is usually system local time). When detecting
23:59:59, the timestamp is modified to represent 0:00
of the next day.
But when writing back the timestamp into the field
(usually DTEND), the now changed time zone information
was not written along with it.

So for vCalendar 1.0 that use UTC representation of
local time 23:59:59, the value written back to
DTEND was UTC 0:00 instead of localtime 0:00.

For vCalendar 1.0 records which use floating
(i.e. localtime) representation of DTEND for
allday events the code worked because no
time zone conversion took place.