vtimezone: accept alternative line ends, using const string & for efficiency
1. timezone parsing: accept \r\n and \r as line ends in addition to \n

At the core of the old code was a string search for BEGIN:<param>\n.
This only worked for \n line ends, for other line ends the search
gave no results.

The new FindProperty() and RfindProperty() search functions accept
the same search string as before (with \n line end), but match that
character against \n, \r (could be part of \r\n) and the end of the
string (not a problem yet, but at some point the last line might not
have a line end character).

They fall back to normal string searches if the search string doesn't
end in \n. A single \n is also handled and does a search for any kind
of line break.

2. vtimezone: use const string & instead of string copy

Change argument parsing so that const references instead of
copies are exchanged. That is more efficient.