DBAPI_TUNNEL_SUPPORT: Implemented now. This allows directly accessing data items from the main application (e.g. for test)
if DBAPI_TUNNEL_SUPPORT is defined in target_options.h,
the engine allows the application to start a pseudo-session
that allows accessing any sync-enabled datastore
through the engine. The data takes the same way as
when read or written during a SyncML session. At the API
level, items are presented in the internal "item" format
based on the configured <fieldlist>. These items can be
accessed on a field level using GetValue/SetValue or
in a serialized key:value text format. These are exactly
the same mechanisms as used in the DB plugin API.

A "tunnel session" always accesses only one datastore.
It is started by passing SESSIONSEL_DBAPI_TUNNEL as
aSelector argument and a datastore name as aSessionName
to OpenSession().
       
The resulting SessionH can then be used to call the
StartDataRead, ReadNextItem(AsKey), ReadItem(asKey),
EndDataRead, StartDataWrite, InsertItem(AsKey),
UpdateItem(AsKey), DeleteItem.
The allowed usage of these routines is the same
as described for a DB API plugin in the SDK manual,
i.e. StartDataRead() must be first, and before
any writing operation StartDataWrite() must be
called, etc.

For the routines that require a aItemKey passed in,
the session key (obtained by OpenSessionKey) can
be queried for it's "tunnel/item" subkey.

The "tunnel/itemtype" value can be read and written
to determine the datatype used for the item (it
defaults to the datastore's preferred send type).