Client: made silently swallow garbage after end of incoming SyncML message
Some implementations send trailing garbage, such as an extra CR, LF, etc.
Until now, the fact that the incoming and outgoing messages share a buffer
in the SyncML Toolkit instance, such garbage was echoed back to the
sender as the first bytes in the next outgoing message.

Now we set the buffer read pointer to the point where the outgoing
message begins in the buffer (where it normally should already be,
but is not in case of trailing garbage).