client state engine: allow multiple calls with STEP_SENTDATA
Without this patch, calling the engine twice with STEP_SENTDATA
results in an error the second time. The second step is redundant,
but shouldn't hurt either. This patch changes the engine so
that it accepts the second (and all further) STEP_SENTDATA by
continueing to ask for data.

This patch simplifies the flow control inside SyncEvolution.