engine: statistics workaround for servers not sending status 201 for completed adds
With the new improved changelog clients now reliably know which
items they send are new and which ones are modified, and
will use <Add> and <Replace> accordingly.

In consequence, receiving ANY ok status for an <Add> can be
safely counted as "remote added" (we KNOW that item is new
in the sync set), even if the server wrongly returns for
example 200 instead of the correct 201 (as Google does).