SAN: moved to SDK, together with the corresponding utility classes
The SanPackage class is meant to be used by clients or utilities
on the server side, not so much by the Engine itself. Putting
the class in the SDK makes it available for those purposes. At
least on Linux, code inside the Synthesis engine cannot be
called unless it is explicitly exported.

This patch, once accepted by Synthesis, relicenses the code
which was moved from sysync to sysync_SDK under the BSD license
of the SDK and thus can be linked statically into apps.

The code has one optional feature: support for SAN messages in
SyncML 1.1 format. This code is compiled unless WITHOUT_SAN_1_1
is defined. Further patches are likely to be needed to get this
compiled, and the application then also must be linked against
the SyncML Toolkit.

The build on Linux disables this feature via config.h, to avoid the
dependency of the SDK on libsmltk - could be made dependent
on a configure option.