server engine prefix: same as in client engine
The SDK only supports connecting to libraries which provide
SySync_ConnectEngine or ConnectEngine. Therefore the engine
cannot be compiled so that it provides SySyncSrv_ConnectEngine,
even though that would be desirable. The SDK needs to be adapted
first (allow SDK user to choose between client and server, pick
right prefix accordingly), then this patch can be reverted.