Linux platform_DLL: open shared library with name as requested
On Linux it may be necessary to open a library with a name
like libfoo.so.2: this is the library libfoo.so with the API
revision 2. The corresponding libfoo.so is normally only used
when linking an application and might not be installed at
runtime.

With this patch, opening such a libfoo.so.2 becomes possible.
Previously the .so suffix was always appended.