SDK engine connect: allow choosing between client and server engine
This patch extends the semantic of the "name" parameter in ConnectEngine()
so that it becomes possible to choose between the client and server
version of the engine when both are inside the same library or
executable (linked statically).

The semantic now is:
        "[]" = builtin client engine
        "[server:]" = builtin server engine (different names of entry points)
        "<foo>" = client engine inside dynamic library <foo>
        "server:<foo>" = server engine inside dynamic library <foo>

The server engine has to use the "SySync_src_" prefix for its entry points
when it is to be opened with the "server:" qualifier. If it uses the normal
"SySync_" prefix, then it can be opened without that qualifier, as before.

Support for "[server:]" depends on a new define, DBAPI_SRV_LINKED. This
was made conditional so that users of the SDK are not forced to link
against the server engine (similar to DBAPI_LINKED for the client engine).

At the moment, DBAPI_SRV_LINKED is not set anywhere. This will be added
for compilation via autotools in a separate commit.