autotools build: added --with-syncml-engines, allow linking both
The new --with-syncml-engines=client|server|both|none allows
to choose which engines are compiled. Currently this results
in libsynthesis.so (as before) and libsynthesisserver.so (the
server engine).

The server version uses different names for all its symbols
because the "namespace sysync" is renamed to "namespace sysyncsrv"
via a -Dsysync=sysyncsrv in the CPPFLAGS. This allows linking
both into the same executable. Several other name clashes
had to be removed by avoiding extern "C" linkage (see previous
patches).

The SDK is compiled so that it can use these directly linked
engines via "[]" and "[server:]" *only* if everything is build
statically (--disable-shared). In that case it wouldn't be
possible to open the engines dynamically. The --with-syncml-engines
choice is taken into account, so that there are no undefined
refences to entry points of engines which were disabled.

When building shared libraries, the SDK has no direct reference
to engine entry points and thus no longer supports "[]".
This allows users of the SDK to choose the engine without being
forced to bring both into memory at the same time (although that
would work).

--disable-shared is useful for development, whereas --enable-shared
is recommended for production binaries. Apps need to be aware
of the difference and call ConnectEngine() differently. The synthesis.pc
file has *not* been updated to cover this difference. It should be
good enough for SyncEvolution, but more work would be needed: either
merge client and server engine into the same library (Lukas
already started with this) and/or rethink what synthesis.pc is really
meant to provide to its users.