unilib: binfile/customimpl: Now binfile can be included but kept inactive
This is necessary as in unilib we need the binfile layer for clients,
but can't use it for servers (binfile layer implementation is capable
for client only at this time). So in server case it is present, but
switched off.

Pure clients can be compiled with BINFILE_ALWAYS_ACTIVE, which
leaves out the code that is required to run customimpl in both
modes (and quite a bit of code in customimpl needed for admin).