autotools build: switched to new combi engine
configure/make build all files required for client
and server engine into the same libsynthesis. The
SDK always assumes that the engine is linked
directly and thus has undefined references to the
engine entry points.

An application which doesn't want to link against
the engine has to provide dummy entry points, then
later open libsynthesis.so.0 in ConnectEngine. Should
be provided as part of the Synthesis source code, together
with a corresponding synthesis_sdk.pc.

That the SDK depends on the engine entry points is defined
in config.h, not in the shared SDK/target_options.h.
The rationale is that not all SDK users might want the
extra dependency. To make this work, the relevant SDK
source code has to include config.h when available.

This patch builds upon the (incompletely merged and obsolete)
build update for two separate libraries and thus removes
obsolete code. It cleans up the selection of files a bit
(SDK contained some unused code).