64 bit: adjusted generic_types.h to make sure uInt32/sInt32 IS 32bits
(I am compiling under Snow Leopard 10.6, which by default creates
64-bit debug builds. Deployment builds are still 10.4/32-bit)

So far, a trick was in use which defined uInt32 such that
it had 64bit apparently to avoid some issues.
However, that caused other issues, like MD5 not working any
more (took me several hours to track down a server auth
error to this - thought is was my unilib changes).

Anyway, this is UGLY - xxx32 must be 32 bits under all
circumstances IMHO, or it is totally misleading.

Steps to improve the situation:

- uInt32/sInt32 now is really 32bit

- new types uIntArch/sIntArch are defined to designate
  "architecture width integers" (but always 32bits or
  more).

- UINT32_HAS_64BITS_ON_ARCH64 can be defined (now
  commented out) to get back to the situation we
  had until now.

- used uIntArch in platform_thread.h now.

- adapted platform_thread.cpp in unix_common
  to be 64bit compatible (using uIntArch).
  Is ok for 32/64bit Mac, should be as well for Linux.

- Codewarrior is 32bit only anyway, so uIntArch
  remains 32bit and implementation does not need
  to change. Same for other non-64bit capable
  platforms.

- Some uInt32 replaced by uIntArch where appropriate
  in other files. Note that generally %ld printfs
  should be fed xIntArch, not xInt32, because "l"
  means long and is 64bit on 64bit architecture.

- pointer that is used in sessionID generation is 64bit
  on 64bit architectures, added needed casts.

The uInt32 being 64 bit issue surfaced with md5, which did
return wrong results. Added extra safeguarding in MD5 to make
sure it is stable in all cases:

- defines own types now, as in current 64bit builds
  MD5 was not working because SYSYNC_UINT4 could
  get 64bit (with UINT32_HAS_64BITS_ON_ARCH64 set
  in generic_types.h)

- Added md5::dotest() in syncappbase when
  MD5_TEST_FUNCS is defined.