syncappbasse: turned extern "C" into static functions
The SyncML Toolkit and expat callbacks don't need to be externally
visible. Making them static reduces the symbol table that the
linker has to deal with and (more imporantly) avoids name clashes.

The callbacks are still defined inside an 'extern "C"' block,
so if that has any effect, it still should apply.