autotools + pkg-config: added synthesis-sdk.pc
In contrast to synthesis.pc, synthesis-sdk.pc does not link
against the engine. Applications which want to install and
load the engine as module can use synthesis-sdk.pc.

It provides the engine entry points expected by libsynthesissdk.a
in a dummy library (libsynthesisstubs.a), so that linking
against the SDK succeeds without undefined symbols.

From an API perspective, the difference between the two methods
is that synthesis.pc supports "[]" and "[server:]" whereas trying
that with synthesis-sdk.pc will fail.