vcard/vcal B64 properties: base64 encoded properties must be terminated with a CRLF even if folding is disabled
The terminating CRLF is automatically added as part of the
encodeValues()/finalizeProperty() process, but only if
folding is enabled.

Because the terminating empty line is a irregularity
to the folding process, I decided to add this special
case (no folding, but still needing ONE linebreak at
the end) here rather than modifying the normal way
encodeValues()/finalizeProperty() work (the former
is also used from elsewhere in the code).