vcard/vcal folding: Made sure folding does not occur within UTF-8 sequences
While the specs are not entirely clear about this,
and it does not matter when unfolding preceedes any
interpretation as a character string, folding within
a UTF-8 sequence makes the vCard/vCalendar as whole
a non-string, which for example cannot be assigned
to a Mac OS X NSString (because these strings must
have valid encoding in all cases).

So this patch adds a check which make sure folding
only occurs at UTF-8 boundaries.