CRC change detection: added <crcchangedetection> to enable CRC-based change detection
<crcchangedetection> is available in binfile based
builds with CHANGEDETECTION_AVAILABLE to have the engine
ignore the modified status delivered by the database
layer (if any) and detect changes based on CRC
checksums.

<crcchangedetection> is off by default; enabling it
causes customimplds to always request all records
from the DB layer (as in slowsync or with dynamic
filters). However, it still does not require the
DB layer to provide data at ReadNextItemXX() - if
the DB layer does not provide the data there, the
engine will read each item explictly.