Unified "undefined field" ID definition - FID_NOT_SUPPORTED and VARIDX_UNDEFINED have the same meaning
This could cause problems when looking for fields for filtering,
because the involved routines used both symbols but checked for
equal with one of both at some places.

FIXME: The two defines exist for historical reasons, and should be eliminated
in a future cleanup. For now, a #ifdef-#error makes sure they have the same
value (which must be in 8-bit range when used with scripts).