TSyncSession::checkCredentials(): <requiredauth>none => accept invalid credentials
This changes the meaning of <requiredauth>none</requiredauth> from "all kinds of authentication
methods allowed, including none" to "all kinds of credentials allowed, including invalid ones".
Because a client with invalid credentials was already allowed to connect by not sending them,
this doesn't make a server setup less secure. It has the advantage that the client setup becomes
easier.