TEngineInterface::debugPuts(): not active in server
The real implementation of TEngineInterface::debugPuts()
was unintentionally depending on the DBAPI_TUNNEL_SUPPORT
define. In the server engine, DBAPI_TUNNEL_SUPPORT wasn't set,
so the default implementation which only returned an error
was active.

Now client and server implement the call. The downcast to
TSyncClientBase was unnecessary, the base class also provides
getDbgLogger().