client side filtering: fixed to work with dynamic filters as well
So far, only static filtering was in use (in TodoZ).
Dynamic filtering was broken in that it caused all
records that passed the filter to be sent to the
server in every sync, because the code did not
expect implGetItem to return the non-changed items
as well (but for dynamic filters, it has to!).

Now client and server case are much more similar.

Furthermore, added implTracksSyncopChanges() so
stdlogicds can know if the database can track
what items belong to the sync set in the case
the items in question have been excluded from
the syncset by a dynamic filter (and reported
back to the DB via reviewReadItem()).

Binfile based datastores cannot track these,
but ODBC/SQL/DBApi based ones can.