log file base name: configurable via "filename" config option
Previously, the base part of the log file name was hard-coded inside
libsynthesis at compile time via the TARGETID. Apps might have no
control over how the library is compiled, so this patch adds a
configuration option called "filename" (similar to
"filesuffix/prefix") for this.

SyncEvolution uses this option to ensure that the filename is
"syncevolution-log.html", no matter how the library was compiled.
It must know what the file is called because it prints the name
in an error message on the command line.