SDK_util.c: fixed crash in vasprintf()
When the vasprintf() fallback is invoked for "va_list args",
that value was already used and modified by vnsprintf(). va_lists
have to be copied when using them multiple times.

Not doing so leads to reading the wrong values from the stack and
(in the case of a %s format) accessing random memory.