Linux: improved support for static linking via //static/
On Linux/Unix, statically linking plugins was possible only via
dlsym(RTLD_DEFAULT) lookups. This depended on --export-dynamic or
specific linker scripts. Both RTLD_DEFAULT and --export-dynamic are
not portable.

This patch adds a different approach to solve this problem: the plugin
has to be specified using the format //static/sym=addr/... where sym
is a name of a function (as asked for by libsynthesis, like Module_Capabilities)
and addr a decimal value giving the address *in the running process*.

This syntax obviously depends on creating the XML config and its
<plugin_module> value at runtime because predicting these values is
hard. On the other hand, it is more portable.