autotools: added more configure options
All optional library dependencies can now be disabled via
--disable switches. The default is that available libraries
are used, as before.

The advantage is that a simpler libsynthesis can be built on
a build machine which has the development files for these
optional libraries installed.

An explit "--enable" has no effect, the library is only used
when it is really available.