debuglogger: avoid the need to create an empty va_list (which is not cleanly possible)
just passing a va_list stack var causes gcc to complain
it was not initialized.

But apparently, the only way to init a va_list is va_start,
which in turn can only be used in functions which actually
have a ... in the declaration.

So instead messing with va_xxx, we now just call another
overloaded variant of DebugOpenBlock().