do not treat 418 = "item exists" as an error
418 = item already exits: sent by Funambol server when both client and
and server have a new item which is considered identical by the server
(must be really identical, minor difference will lead to a merged item
which is sent back to the client without the 418).

This patch treats 418 as an "okay" status code and lets the session
continue.

Found in SyncEvolution, see Moblin Bugzilla #4599.