implemented superdatastore support in binfile clients
Several different problems prevented using of <superdatastores> in binfile
clients. This patch fixes most, but not all of them. With this patch,
a <superdatastore> can be used like this:
- It does not have its own target in the client's profile.
- Progress events are produced for the sub-datastores.
- The sub-datastores must be configured via their targets so
  that they have the same sync mode and the special URI <super:uri>
  (including the angle brackets!), where "super" is the name
  of the <superdatastore> and "uri" its URI on the server.

The most obvious problem with the previous code were crashes when
TBinfileImplClient::SelectProfile() did static casts into TBinfileDSConfig
and TBinfileImplDS. The config and implementation of <superdatastore> are
not derived from these classes. With this patch, such upcasts are done
via virtual methods that return NULL when the case is impossible. This
is a substitute for dynamic_cast<>.

For TBinfileImplDS, this additional check is not strictly necessary at the
moment, but is done anyway to make the code safer.

TSuperDataStore creates TBinfileImplDS
instances. TBinfileClientConfig::newProfile() must be aware of that
and avoid creating them once more, which changes the logic of when a
store is added to fLocalDataStores.

The configuration of the TSuperDataStore is taken from the first
TBinfileDSConfig which triggers its creation. This moves around some
code. Static utility code in TLocalEngineDS helps parsing the special
URI format.

Order of the normal datastores and TSuperDataStore matters while
calling the engPrepareClientSyncAlert() methods. The TSuperDataStore
must be called last, which is achieved by putting them at the end of
fLocalDataStores once all stores are created.

TLocalEngineDS::engProcessSyncAlert() was too strict with its check
of just fAsSubDatastoreOf: as the comment said, it should detect
that the store is alerted again as normal store, which implies that
the if() must look at the aAsSubDatastoreOf parameter. In a client,
fAsSubDatastoreOf is set in engPrepareClientSyncAlert() (after
searching for the superdatastore in TSyncAgent::NextMessage()),
then later engProcessSyncAlert() is called again with the same,
non-NULL aAsSubDatastoreOf by TSuperDataStore.

Setting fAsSubDatastoreOf early via engPrepareClientSyncAlert() is
necessary to prevent generating separate <Alert>s for the
sub-datastores. However, this leads to a problem when the same engine
is used twice because of an authentication failure: the second time,
engPrepareClientSyncAlert() complains because fAsSubDatastoreOf is
still set from the first time and aborts.  Perhaps the TSuperDataStore
parameter of engPrepareClientSyncAlert() can be removed when changing
the check for it in engGenerateClientSyncAlert()?

TLocalEngineDS::engInitForSyncOps() didn't work in a client (detected
by checking the special URI) and didn't seem to be necessary. With
this patch, it is a NOP in clients.

Fixed a cut-and-paste error in TSuperDataStore::engGenerateMapItems():
didn't use the "pos" iterator of the loop in that method.