unconfirmed map + superdatastore: fixed mismatch between local ID with and without prefix
TLocalEngineDS::engGenerateMapItems() stored IDs with the prefix given
to it by TSuperDataStore in fUnconfirmedMaps. Later the
TSuperDataStore tried to remove them without the prefix, which failed
and lead to an incorrect list of unconfirmed mappings.

This patch assumes that the behavior of storing IDs without the prefix
inside the normal datastore is the desired one and thus fixes
TLocalEngineDS::engGenerateMapItems(). It now works for normal syncs;
suspended syncs were not tested.