DevInf + superdatastore: avoid redundant <DataStore>
When the client has alerted a superdatastore in a SyncML server
and then is sent DevInf, the server produced a <DataStore>
entry <DevInf> for the superdatastore and each sub-datastore,
all with the same <LocURL> (the one that was alerted).

This is redundant at best, and confusing the client at worst.
This patch detects that a datastore was alerted as a sub-datastore
and skips it when generating <DevInf>.