debuglogger: avoid segfault when log file cannot be reopened
While a SyncML server session was active, I removed the directory
and the log file that the server was using. The next time the
server tried to log something, it segfaulted in the fclose(fFile)
because of the NULL pointer from fopen().

Don't do anything if the file cannot be opened.