server engine: added missing check for re-sent <Add> during resume.
This remained undetected because most clients, including
libsynthesis itself before a few months ago, always used
<Replace> and never <Add> (which is ok for a client to
do).

Now an <Add> with an already known remoteID will be
detected and causes implProcessItem() to return
status 418.
This status will be either reported back to the client
in a resumed session, or in a non-resumed session
with updates not explicitly ignored the operation
will be implicitly converted to a <Replace>. In the
server case, 208 status will be returned.

Note that for the client case, duplicate adds are
detected and prevented entirely differently (by
checking against list of pending maps = IDs of
not completely confirmed and mapped adds from the
previous session).