TBlob: export Init() properties read-only
Making these properties available for a user of TBlob is useful
because then the user can computer some values, call Init() and
later retrieve them without having to store its own copy.

SyncEvolution does that with the blob path. The other properties
are added in this patch for the sake of completeness.