resume + pending item fix in SyncML server
When an item that was split among messages was finally reassembled
and processed by the server, the status sent to the client was not
properly stored in the admin data. As a result of that, when the
session got suspended directly after sending that status, the client
would resume, send the last item chunk again, but then not get
the previous status. Instead the server attempted to parse the
incomplete item and failed.

The root cause was that TSyncSession::process() was called
recursively. The inner one for the incomplete command saved the
status in fLastItemStatus. The outer one then later didn't have
a status pointer, fell back to status 0 and overwrote the valid
status.

This patch solves this by suppressing that second updating of
fLastItemStatus and all related fields if the item was handled
by an inner process() call.