engine: getItemByID: made sure newly inserted items can be retrieved
getItemByID assumed that all items to be read were part
of the syncset which is obviously not the case for
newly inserted items.
So in case the item is not found in the sync set, we
must try to directly load it from the DB.

This bug only affects customimpl on binfile AND
change detection by CRC enabled. No product before
the Android client used this combination so far
so it did not surface until now.