engine: fix for very late suspend
When a client was requested to suspend, datastores that
had already successfully finished syncing were still
considered "aborted" and thus did not save their
anchors properly.

With this change a late suspend will only affect
datastores that were still in progress at the
time of suspend. Already completed datastores
will save anchors and will NOT resume in the
next sync.