TZ init: split initialisation of system time zone list into two parts (basic usability of TZ system must be ready before config reading, extra TZ retrieved from system can be added later)
Previously, all parsing of system time zone information was done in the
TSyncAppBase constructor. This had two problems:
1. failure to parse couldn't be reported to app (as mentioned in a
   TODO)
2. logging in this fairly complex code was impossible, because
   the logger was not initialized yet

This patch adds a finishConfig() callback and uses it to invoke
a second global time zone initialisation function. On Linux, that function is used
to add the libical time zone definitions and logs them at the same time.
Builtin time zones are still added at appBase construction time, before parsing the
config, such that a basic TZ environment is available already during
config parsing.
TZ that only get added in the second init run of course cannot be referenced
in the config file (but that is unlikely to be required at all).