TSyncAppBase: moved system time zone parsing after config reading
Previously, fAppZones.initialize() and thus parsing of system time
zone information was done in the TSyncAppBase constructor. This had
two problems:
1. failure to parse couldn't be reported to app (as mentioned in a
   TODO)
2. logging in this fairly complex code was impossible, because
   the logger was not initialized yet

This patch moves the initialize() call after config reading. There are
two such places, one for a real XML config and one for a hard-coded one.
The common code is in the new finishConfig() method, which is only
invoked if there was no previous error and then provides the final
return code for config reading. Like some other code it sets the error
in the config as well as returning a LOCERR_CFGREAD.